Move environment parsing tests out of the unit test suite
… since they can influence other tests.
